+++ packages/binding/src/services/completion/providers/create-key-value.ts
@@ -5,9 +5,8 @@ import {
 } from "vscode-languageserver-types";
 import { BindingParserTypes as BindingTypes } from "@ui5-language-assistant/binding-parser";
 
-import { propertyBindingInfoElements } from "../../../definition/definition";
+import { getPropertyBindingInfoElements } from "../../../definition/definition";
 import { typesToValue } from "../../../utils";
-import { getDocumentation } from "./documentation";
 import { BindContext } from "../../../types";
 
 export const createKeyValue = (
@@ -15,7 +14,7 @@ export const createKeyValue = (
   binding: BindingTypes.StructureValue
 ): CompletionItem[] => {
   // exclude duplicate
-  return propertyBindingInfoElements
+  return getPropertyBindingInfoElements(context)
     .filter((item) => {
       if (
         !binding.elements.find(
@@ -29,13 +28,12 @@ export const createKeyValue = (
     .map((item) => {
       const type = typesToValue(item.type, context, 0);
       const text = `${item.name}: ${type.length === 1 ? type[0] : "$0"}`;
-      const documentation = getDocumentation(item);
       return {
         label: item.name,
         insertTextFormat: InsertTextFormat.Snippet,
         insertText: text,
         kind: CompletionItemKind.Snippet,
-        documentation,
+        documentation: item.documentation,
       };
     });
 };